requirements.txt
python==3.8
biopython==1.79
joblib==1.1.0
numpy==1.23.1
pandas==1.4.3
pymongo==3.12.0
PyYAML==6.0.2
scikit-learn==1.1.1
scipy==1.8.1
notebook==6.5.6
